requirements-agent.txt
gspread>=6.0.0
google-auth>=2.0.0
watchdog>=4.0.0
requests
